Fixed Content-Length deletion for reverse proxies in Martini
It turns out that deleting Content-Length header field at the end of
serveGzip function may not be enough for gzip. According to the spec
of compress/gzip, a Write() call may flush the writer before closing
it. This means that in certain cases, gzip writer flushes compressed
data to the underlying martini.ResponseWriter with a wrong header
Content-Length field, i.e before it is deleted by serveGzip resulting
in a "Conn.Write wrote more than the declared Content-Length" error.
This change adds a defensive measure of deleting Content-Length
before a potential write.